Fix error bundling and Delete report structure
