+++ model/model.go
@@ -15,6 +15,10 @@
  */
 package model
 
+import (
+	"log"
+)
+
 type ShutdownReport struct {
 	// InstanceGroup, ComputeEngine, SQL
 	InstanceType string
@@ -22,4 +26,29 @@ type ShutdownReport struct {
 	DoneResources []string
 	// already stopped resource names
 	AlreadyShutdownResources []string
+}
+
+const (
+	ComputeEngine = "ComputeEngine"
+	InstanceGroup = "InstanceGroup"
+	SQL           = "SQL"
+)
+
+func (r *ShutdownReport) Show() {
+	log.Println("<<<<< " + r.InstanceType + " >>>>>")
+	if r == nil {
+		log.Printf("There are no instances in %s. Skip.\n", r.InstanceType)
+		return
+	}
+	log.Println("!REPORT!")
+	log.Println("[Shutdown Resource]")
+
+	for i, resource := range r.DoneResources {
+		log.Printf(">> Resouce(%v): %v\n", i+1, resource)
+	}
+
+	log.Println("[Already Shutdown Resource]")
+	for i, resource := range r.AlreadyShutdownResources {
+		log.Printf(">> Resouce(%v): %v\n", i+1, resource)
+	}
 }
