+++ operator/instance_group.go
@@ -22,7 +22,6 @@ import (
 
 	set "github.com/deckarep/golang-set"
 	"github.com/future-architect/gcp-instance-scheduler/model"
-	"github.com/future-architect/gcp-instance-scheduler/report"
 	"github.com/hashicorp/go-multierror"
 	"golang.org/x/net/context"
 	"google.golang.org/api/compute/v1"
@@ -147,7 +146,11 @@ func (r *InstanceGroupListCall) FilterLabel(targetLabel string, flag bool) *Inst
 	}
 }
 
-func (r *InstanceGroupShutdownCall) ShutdownWithInterval(ctx context.Context, interval time.Duration) (*report.ShutdownReport, error) {
+func (r *InstanceGroupShutdownCall) ShutdownWithInterval(ctx context.Context, interval time.Duration) (*model.ShutdownReport, error) {
+	if r.Error != nil {
+		return nil, r.Error
+	}
+
 	var res = r.Error
 	var doneRes []string
 	var alreadyRes []string
@@ -160,11 +163,6 @@ func (r *InstanceGroupShutdownCall) ShutdownWithInterval(ctx context.Context, in
 
 	// instance group manager service
 	ms := compute.NewInstanceGroupManagersService(s)
-
-	// error bundle before executing stop call
-	if res != nil {
-		return nil, res
-	}
 
 	for _, manager := range valuesIG(r.InstanceGroupList.Items) {
 		// get manager zone name
@@ -204,11 +202,9 @@ func (r *InstanceGroupShutdownCall) ShutdownWithInterval(ctx context.Context, in
 	}
 	log.Printf("Success in stopping InstanceGroup: Done.")
 
-	return &report.ShutdownReport{
-		ShutdownReport: model.ShutdownReport{
-			InstanceType:             report.InstanceGroup,
-			DoneResources:            doneRes,
-			AlreadyShutdownResources: alreadyRes,
-		},
+	return &model.ShutdownReport{
+		InstanceType:             model.InstanceGroup,
+		DoneResources:            doneRes,
+		AlreadyShutdownResources: alreadyRes,
 	}, res
 }
